raw conversion of StrictFake specs
